added Dagger

+++ app/src/main/java/com/example/weathercomposeapp/domain/repository/WeatherRepository.kt
@@ -5,8 +5,8 @@ import com.example.weathercomposeapp.domain.entity.Weather
 
 interface WeatherRepository {
 
-    suspend fun getWeather(cityId: Int): Weather
+    suspend fun getWeather(cityId: Long): Weather
 
-    suspend fun getForecast(cityId: Int): Forecast
+    suspend fun getForecast(cityId: Long): Forecast
 
 }
